+++ CHANGELOG.md
@@ -5,10 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.1.8.4]
+## [0.1.9]
+### Added
+ - Network graph analysis
+
+### Changed
+ - Use napari's built-in progress bar
+ - Contrast analysis/PLS analysis run as async to keep GUI alive
+ - Use napari's new text overlay in analysis step
 
 ### Fixed
  - Fix tensorflow version for StarDist
+ - Fix project unable to save after elastix
 
 ## [0.1.8.3]
 
@@ -34,6 +42,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### TBD
 
 [unreleased]: https://github.com/olivierlacan/keep-a-changelog/compare/v1.1.1...HEAD
+[0.1.9]: https://github.com/olivierlacan/keep-a-changelog/compare/v0.1.8.3...v0.1.9
 [0.1.8.3]: https://github.com/olivierlacan/keep-a-changelog/compare/v0.1.8.2...v0.1.8.3
 [0.1.8.2]: https://github.com/olivierlacan/keep-a-changelog/compare/v0.1.8...v0.1.8.2
 [0.1.8]: https://github.com/olivierlacan/keep-a-changelog/compare/v0.1.7...v0.1.8